+++ osidb/__init__.py
@@ -1,5 +1,5 @@
 """
     osidb version
 """
-__version__ = "3.5.1"
+__version__ = "3.5.2"
 VERSION = __version__