+++ scripts/run-celery-standalone.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 
 rm -f /tmp/celery_worker.pid
-exec celery -A config worker --pidfile /tmp/celery_worker.pid -f celery.log --loglevel DEBUG -P gevent --concurrency=1 -Q slow,fast -E
+exec celery -A config worker --pidfile /tmp/celery_worker.pid -f celery.log --loglevel DEBUG --concurrency=3 -Q slow,fast -E